Issue #14: Make cluster version available
Motivation
----------
To use ALTER INDEX statements we must ensure that the cluster version is
at least 5.5.

Modifications
-------------
Use /default/pools to get the list of nodes and select the lowest
clusterCompatability to determine the effective cluster version.

Pass this value through to the index definition to determine mutations.

Also use this value instead of the --no-rbac flag to recognize we're on
a 4.0 cluster and force manual replica management.

Results
-------
Cluster version is now available for use for replica management in
future commits.

Additionally, we are now compatible with 5.0 clusters that are still
running with authentication disabled.