Issue #14: Update index node assignments
Motivation
----------
Support changes to index node assignments after an index is created.

Modifications
-------------
Implemented a phase system for mutations so that different replicas can
be moved separately to prevent downtime.

Fixed a flaw in the logic using the /indexStatus API to retrieve node
assignments for automatic replicas.

Created MoveIndexMutation to handle ALTER INDEX statements on 5.5, with
support in IndexManager.

Created a FeatureVersion class to help manage Couchbase versions where
specific features are supported.

Return proper mutations for changes to nodes and num_replicas, in both
the manual and automatic index mutation forms.

Change UpdateIndexMutation to recognize when it's actually a safe
mutation based on replicas being present.

Results
-------
Automatic replica scaling is supported on CB >= 5.0 as an unsafe
operation.

Automatic replica moving is supported on CB >= 5.5 as a safe operation,
moves are skipped with a warning on 5.0 and 5.1.

Manual replica scaling and moving is supported, and is safe so long as
there is at least one replica.

Other update operations, such as index_key and condition, are now safe
operations for any index with at least one replica and running in manual
replica mode.