Issue #32: Output replica count when plan is reducing to 0
Motivation
----------
When reducing the number of replicas to zero, the Update statement in
the plan doesn't make clear what the change will be.

Modifications
-------------
Include the replica count in the Update statement of plans if either the
existing or new replica count is non-zero.

Results
-------
Operators can more readily see what changes their plan is making.