Upgrade to Couchbase SDK 3.x
Motivation
----------
Modernize.

Modifications
-------------
Upgraded to Couchbase SDK 3.0.4 and rewrote the ConnectionManager and
IndexManager to work with the new SDK.

Removed the --no-rbac option, since it is no longer required to support
authentication on 4.x when using the new SDK.

Upgraded babel and eslint.

Upgraded Docker image to use Node 12 (current LTS).