Add support for num_partition
Motivation
----------
Allow the number of partitions in a partitioned index to be manually
controlled.

Modifications
-------------
Added an option to the partition definition to control the number of
partitions. If missing, the number is ignored. If present, the index
is synchronized to have that number of partitions.

Fixed a bug in SDK 3.x upgrade that prevented partitioned indexes from
working as expected.