Add support for ALTER INDEX replica_count
Motivation
----------
Allow the number of replicas for an existing index to be changed without
downtime using the ALTER INDEX statement on Couchbase Server 6.5.

Modifications
-------------
Add ResizeIndexMutation and wire it up to be used whenever the Couchbase
server is on at least version 6.5.

Results
-------
Safe changes with no index downtime are performed if the number of
replicas is changed and the server is on 6.5 of later. No changes to
behavior for older Couchbase Server versions.