Avoid name conflicts with scope/collections (#79)
Motivation
------------
Scopes and collections allow indexes with the same name across different
collections.

Modifications
---------------
Normalize the API responses for scope/collection information, and when
matching an index ensure we account for the scope/collection.

When building indices or waiting for the build to complete, target a specific
scope/collection (or default).

Add travel-sample to the dev environment since it has scopes/collections.

Results
--------
couchbase-index-manager is now compatible with scopes/collections in Couchbase
Server 7.0. The presense of indices within a collectdion will no longer confuse it.

However, it isn't yet capable of managing indices in a collection.